Added products.test.js and removed product tests from customers.test.js
